Random objects - inportant
Instead, to repeat the random lines of code for any attribute inside the which object,we put it out of the object and assign one attribute to put inside the object.We could try to do it inside with any animals class but doing like that wound not be possible to add animals as a hard code to keep the algorithm more flexible. that's why we follow that way.